requirements.txt
flask
openai
python-dotenv
pyttsx3